requirements.txt
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
Jinja2==3.1.3
kiwisolver==1.4.5
loguru==0.7.2
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
sympy==1.12
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.2
typing_extensions==4.11.0
